% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullsg.R
\name{apply_option_order}
\alias{apply_option_order}
\title{Make survey_responses columns into ordered factors based on survey_questions options.}
\usage{
apply_option_order(survey_responses, survey_questions)
}
\arguments{
\item{survey_responses}{Return value from pullsg}

\item{survey_questions}{Return value from get_questions}
}
\value{
New version of survey_responses, re-leveled
}
\description{
By default, allows
}
