% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullsg.R
\name{get_question_options}
\alias{get_question_options}
\title{Get all the options for all questions with options (e.g., radio, table, checkbox).}
\usage{
get_question_options(question_df)
}
\arguments{
\item{question_df}{Return value of get_questions}
}
\description{
Removes disabled options (question_df$options$properties$disabled FALSE).
}
\details{
Return a data frame with option_id, option_type, title, title_language, value, order,
question_id, question_type, question_subtype, question_qtext.
}
