% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullsg.R
\name{get_questions}
\alias{get_questions}
\title{Fetch the question information, and make up variable names in the qtext column.}
\usage{
get_questions(lc_qurl, default_var_name_policy)
}
\arguments{
\item{lc_qurl}{URL whence to fetch question json}

\item{default_var_name_policy}{See pullsg}
}
\description{
qtext (question name) is whichever is non-NA from shortname, sub_varname,
and the title made up from default_var_name_policy.
}
\details{
Returns a data frame with all columns.
}
