% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullsg_campaign.R
\name{pullsg_campaign}
\alias{pullsg_campaign}
\title{Download survey campaign data from Survey Gizmo, storing an R dataframe.}
\usage{
pullsg_campaign(surveyid, api, reset_row_names = T, customfields = T,
  contactinfo = T, small = F, verbose = T)
}
\arguments{
\item{surveyid}{The survey's unique SG ID number (in V4 of the API, the portion of the \href{https://apihelp.surveygizmo.com/help/article/link/surveyresponse-sub-object}{surveyresponse} call URL which follows "id/", e.g.: "...build/id/1234567"}

\item{api}{The user's private API key for Survey Gizmo}

\item{reset_row_names}{When true (the default), resets row names to 1, 2,..N in the returned dataframe}

\item{customfields}{When true (the default), SG's "custom fields" associated with the contact are preserved in the returned data.}

\item{contactinfo}{When true (the default), SG's phone and physical address fields associated with the contact are preserved in the returned data.}

\item{small}{When true, the only fields returned are contactid and email address.}

\item{verbose}{When true (the default), download progress is printed to standard output.
\strong{NOTE:} When set to false, pullsg will change the stub of SG system variables to \emph{sys_*}}
}
\description{
This function downloads campaign data from Survey Gizmo (SG), saving
the returned email/response data as an R \code{\link{data.frame}}. Because SG limits the size of JSON data pulls via the API (currently, the limit is 250),
it calculates the number of pulls needed to download the entire response set and binds the returned frames. The "contactid" field joins email address and other campaign data with the survey response objects (see \code{\link{pullsg}}).
Note that this function only downloads contact object data that is accessible through the API, specifically, contact objects associated with the subtype "email" where status is "Active."
}

