% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsurveygizmo.R
\docType{package}
\name{Rsurveygizmo}
\alias{Rsurveygizmo}
\alias{Rsurveygizmo-package}
\title{Rsurveygizmo: a method for downloading survey response data from the Survey Gizmo AWS servers}
\description{
\pkg{Rsurveygizmo} allows you to access the Survey Gizmo (SG) REST API
directly from the R console.
By specifying your API key and a survey project ID you are able to download
current SG survey responses into an R \code{\link{data.frame}}.
}
\details{
The only function currently provided by \pkg{Rsurveygizmo} is
\code{\link{pullsg}}, which pulls data from the SG servers and includes a
variety of parameters to control the format of the returned
object (see the \code{\link{pullsg}} help file for details).
}

