% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullsg.R
\name{get_questions}
\alias{get_questions}
\title{Fetch the question information, and make up variable names in the qtext column.}
\usage{
get_questions(
  surveyid,
  api,
  secret,
  default_var_name_policy = c("id", "title", "title_plus_id"),
  locale = "US"
)
}
\arguments{
\item{surveyid}{Survey id}

\item{api}{API token}

\item{secret}{API token secret}

\item{default_var_name_policy}{How to make a variable name.}

\item{locale}{Locale to use for SurveyGizmo API.}
}
\description{
qtext (question name) is whichever is non-NA from shortname, sub_varname,
and the title made up from default_var_name_policy.
}
\details{
Returns a data frame with all columns.
}
