% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullsg.R
\name{pullsg}
\alias{pullsg}
\title{Pull the raw survey response data from Survey Gizmo, storing an unmodified dataframe.}
\usage{
pullsg(sg_surveyid, api, completes = T)
}
\arguments{
\item{sg_surveyid}{The unique Id number of the survey (in V3, the portion of the
build URL which follows "id": ...build/id/1234567}

\item{api}{The user's unique private API key for Survey Gizmo}

\item{completes}{Should the download include partial responses?}
}
\description{
The function automatically calculates the number of JSON data pulls needed to
download a fully formed extract of the survey. It then binds together all
returned data into a single dataframe for subsequent manipulation.
}

